Redesigned Filter Cards to serve as Configuration Cards, fixed inventory interaction bug, fixed Creative Energy Cube bug, added new recipes

+++ src/main/java/mekanism/common/util/MekanismUtils.java
@@ -1155,7 +1155,7 @@ public final class MekanismUtils
 	 */
 	public static String getEnergyDisplay(double energy)
 	{
-		if(energy == Integer.MAX_VALUE)
+		if(energy == Double.MAX_VALUE)
 		{
 			return LangUtils.localize("gui.infinite");
 		}
